feat: Modifications to build scripts to save env vars
body: Several libraries require additional environment variables
in order to compile fully and without error.  The MODULES=false
case does not currently have a way to save and reload these
environment variables, so this is a stab at implementing a
modules style (but light weight) method to store and recall
environment variables.