feat: adding a no modules type support for nco, jasper, and xerces
body: In an effort to support the build of the ufs-s2s-model stack,
a no-modules method of loading the appropriate environment variables
is implemented here.  The environment variables are stored in the
/etc/profile.d directory as <module>-env-vars.sh and can be sourced
to load each of the environment variables.  This is only meant for
container builds where one is not expected to install more than
1 version of each library.