feat: fixed problems with lapack reference in esmf build
body: although the esmf_build I had modified indicated the mkl
version should be used, it pointed to the free version built
in the jedi-stack.  I bailed on using the mkl version and
called out the netlib version for ESMF_LAPACK env var instead.
I also removed the compile of the NCEPLIBS-external library and
had to modify some of the environment definition files.