feat: fix esmf build to use internal lapack/xerces
body: Due to a failure of the ufs-s2s-model to find a specific
blas function, we have decided to use the internal versions of
lapack/blas and xerces for the esmf build.  As a result, we will
not be building the external lapack or xerces libraries for this
model and will rely on esmf to provide those libraries.